[ch35870] Refactor dialogs, move them in distinct components
