+++ src_ts/elements/common-elements/follow-up-components/follow-up-actions/follow-up-actions.ts
@@ -17,10 +17,9 @@ import isEmpty from 'lodash-es/isEmpty';
 import isEqual from 'lodash-es/isEqual';
 import isObject from 'lodash-es/isObject';
 import isArray from 'lodash-es/isArray';
-import every from 'lodash-es/every';
 import omit from 'lodash-es/omit';
 import each from 'lodash-es/each';
-
+import './follop-up-actions-dialog';
 import '../../../data-elements/get-action-points';
 import '../../../data-elements/update-action-points';
 import {tabInputsStyles} from '../../../styles/tab-inputs-styles';
@@ -34,19 +33,13 @@ import CommonMethodsMixin from '../../../mixins/common-methods-mixin';
 import {getEndpoint} from '../../../config/endpoints-controller';
 import TableElementsMixin from '../../../mixins/table-elements-mixin';
 import DateMixin from '../../../mixins/date-mixin';
-import {
-  actionAllowed,
-  getHeadingLabel,
-  getOptionsChoices,
-  readonlyPermission
-} from '../../../mixins/permission-controller';
+import {getHeadingLabel, getOptionsChoices, readonlyPermission} from '../../../mixins/permission-controller';
 import {checkNonField} from '../../../mixins/error-handler';
 import {sendRequest} from '@unicef-polymer/etools-utils/dist/etools-ajax/ajax-request';
-import clone from 'lodash-es/clone';
-import famEndpoints from '../../../config/endpoints';
 import {AnyObject} from '@unicef-polymer/etools-types';
 import {RootState, store} from '../../../../redux/store';
 import {connect} from 'pwa-helpers/connect-mixin';
+import {openDialog} from '@unicef-polymer/etools-utils/dist/dialog.util';
 
 /**
  * @LitElement
@@ -96,133 +89,98 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
         etools-content-panel::part(ecp-content) {
           padding: 0;
         }
-        etools-dropdown.fua-category {
-          --paper-listbox: {
-                max-height: 340px;
-                -ms-overflow-style: auto;
-            };
-        }
-        etools-dropdown.fua-person {
-          --paper-listbox: {
-                max-height: 140px;
-                -ms-overflow-style: auto;
-            };
-        }
-        .checkbox-container {
-          padding-inline-start: 14px;
-          padding-top: 14px;
-        }
-        .input-container etools-button {
-          height: 34px;
-          color: rgba(0, 0, 0, .54);
-          font-weight: 500;
-          z-index: 5;
-          border: 1px solid rgba(0, 0, 0, .54);
-          padding: 6px 13px;
-        }
-        paper-tooltip {
-          --paper-tooltip: {
-            font-size: 12px;
-          }
-        }
         datepicker-lite::part(dp-calendar) {
           position: fixed;
         }
         .launch-icon {
           --iron-icon-height: 16px;
           --iron-icon-width: 16px;
-          }
+        }
       </style>
 
-      <get-action-points .engagementId="${this.engagementId}" @ap-loaded="${({detail}: CustomEvent) => {
-      if (detail?.success) {
-        this.dataItems = detail.data;
-      }
-    }}"></get-action-points>
+      <get-action-points
+        .engagementId="${this.engagementId}"
+        @ap-loaded="${({detail}: CustomEvent) => {
+          if (detail?.success) {
+            this.dataItems = detail.data;
+          }
+        }}"
+      ></get-action-points>
 
-        <update-action-points .engagementId="${this.engagementId}"
-                              .requestData="${this.requestData}"
-                              .errors="${this.errors}"
-                              .actionPoints="${this.dataItems}"
-                              @ap-request-completed="${this._apRequestCompleted}">
-        </update-action-points>
-        <get-partner-data .partnerId="${this.partnerId}"
-            @partner-loaded="${({detail}) => {
-              this.fullPartner = detail;
-            }}">
-        </get-partner-data>
+      <update-action-points
+        .engagementId="${this.engagementId}"
+        .requestData="${this.requestData}"
+        .errors="${this.errors}"
+        .actionPoints="${this.dataItems}"
+        @ap-request-completed="${this._apRequestCompleted}"
+      >
+      </update-action-points>
+      <get-partner-data
+        .partnerId="${this.partnerId}"
+        @partner-loaded="${({detail}) => {
+          this.fullPartner = detail;
+        }}"
+      >
+      </get-partner-data>
 
-        <etools-content-panel panel-title="UNICEF Follow-Up Actions" list>
-            <div slot="panel-btns">
-                <div ?hidden="${!this.canBeChanged}">
-                  <sl-tooltip content="Add">
-                    <etools-icon-button
-                            class="panel-button"
-                            @click="${this._openAddDialog}"
-                            name="add-box">
-                    </etools-icon-button>
-                  </sl-tooltip>
+      <etools-content-panel panel-title="UNICEF Follow-Up Actions" list>
+        <div slot="panel-btns">
+          <div ?hidden="${!this.canBeChanged}">
+            <sl-tooltip content="Add">
+              <etools-icon-button class="panel-button" @click="${this._openAddDialog}" name="add-box">
+              </etools-icon-button>
+            </sl-tooltip>
+          </div>
+        </div>
+
+        <etools-data-table-header no-collapse no-title>
+          <etools-data-table-column class="col-2" field="reference_number" sortable
+            >${getHeadingLabel(this.optionsData, 'reference_number', 'Reference Number #')}</etools-data-table-column
+          >
+          <etools-data-table-column class="col-3" field="ap_category.display_name" sortable
+            >${getHeadingLabel(this.optionsData, 'category', 'Action Point Category')}</etools-data-table-column
+          >
+          <etools-data-table-column class="col-3">Assignee (Section / Office)</etools-data-table-column>
+          <etools-data-table-column class="col-2" field="status" sortable
+            >${getHeadingLabel(this.optionsData, 'status', 'Status')}</etools-data-table-column
+          >
+          <etools-data-table-column class="col-1" field="due_date" sortable
+            >${getHeadingLabel(this.optionsData, 'due_date', 'Due Date')}</etools-data-table-column
+          >
+          <etools-data-table-column class="col-1" field="priority" sortable
+            >${getHeadingLabel(this.optionsData, 'high_priority', 'Priority')}</etools-data-table-column
+          >
+        </etools-data-table-header>
+
+        ${(this.itemsToDisplay || []).map(
+          (item, index) => html`
+            <etools-data-table-row no-collapse>
+              <div slot="row-data" class="layout-horizontal editable-row">
+                <span class="col-data col-2 truncate">
+                  <a href="${item.url}" class="truncate" title="${item.reference_number}" target="_blank">
+                    ${item.reference_number} <etools-icon class="launch-icon" name="launch"></etools-icon>
+                  </a>
+                </span>
+                <span class="col-data col-3 truncate">${item.ap_category?.display_name || '-'}</span>
+                <span class="col-data col-3 truncate">${item.computed_field}</span>
+                <span class="col-data col-2 truncate caps">${item.status}</span>
+                <span class="col-data col-1 truncate">${this.prettyDate(String(item.due_date), '') || '-'}</span>
+                <span class="col-data col-1 truncate caps">${item.priority}</span>
+                <div class="hover-block">
+                  <etools-icon-button
+                    name="content-copy"
+                    @click="${() => this._openCopyDialog(index)}"
+                  ></etools-icon-button>
+                  <etools-icon-button
+                    name="create"
+                    ?hidden="${!this.canBeEdited(item.status)}"
+                    @click="${() => this._openEditDialog(index)}"
+                  ></etools-icon-button>
                 </div>
-            </div>
-
-          <etools-data-table-header no-collapse no-title>
-            <etools-data-table-column class="col-2" field="reference_number" sortable>${getHeadingLabel(
-              this.optionsData,
-              'reference_number',
-              'Reference Number #'
-            )}</etools-data-table-column>
-            <etools-data-table-column class="col-3" field="ap_category.display_name" sortable>${getHeadingLabel(
-              this.optionsData,
-              'category',
-              'Action Point Category'
-            )}</etools-data-table-column>
-            <etools-data-table-column class="col-3">Assignee (Section / Office)</etools-data-table-column>
-            <etools-data-table-column class="col-2" field="status" sortable>${getHeadingLabel(
-              this.optionsData,
-              'status',
-              'Status'
-            )}</etools-data-table-column>
-            <etools-data-table-column class="col-1" field="due_date" sortable>${getHeadingLabel(
-              this.optionsData,
-              'due_date',
-              'Due Date'
-            )}</etools-data-table-column>
-            <etools-data-table-column class="col-1" field="priority" sortable>${getHeadingLabel(
-              this.optionsData,
-              'high_priority',
-              'Priority'
-            )}</etools-data-table-column>
-          </etools-data-table-header>
-
-          ${(this.itemsToDisplay || []).map(
-            (item, index) => html`
-              <etools-data-table-row no-collapse>
-                <div slot="row-data" class="layout-horizontal editable-row">
-                  <span class="col-data col-2 truncate">
-                    <a href="${item.url}" class="truncate" title="${item.reference_number}" target="_blank">
-                      ${item.reference_number} <etools-icon class="launch-icon" name="launch"></etools-icon>
-                    </a>
-                  </span>
-                  <span class="col-data col-3 truncate">${item.ap_category?.display_name || '-'}</span>
-                  <span class="col-data col-3 truncate">${item.computed_field}</span>
-                  <span class="col-data col-2 truncate caps">${item.status}</span>
-                  <span class="col-data col-1 truncate">${this.prettyDate(String(item.due_date), '') || '-'}</span>
-                  <span class="col-data col-1 truncate caps">${item.priority}</span>
-                  <div class="hover-block">
-                    <etools-icon-button
-                      name="content-copy"
-                      @click="${() => this._openCopyDialog(index)}"
-                    ></etools-icon-button>
-                    <etools-icon-button
-                      name="create"
-                      ?hidden="${!this.canBeEdited(item.status)}"
-                      @click="${() => this._openEditDialog(index)}"
-                    ></etools-icon-button>
-                  </div>
-                </div>
-              </etools-data-table-row>
-            `
-          )}
+              </div>
+            </etools-data-table-row>
+          `
+        )}
         <etools-data-table-row no-collapse ?hidden="${this.itemsToDisplay?.length}">
           <div slot="row-data" class="layout-horizontal editable-row">
             <span class="col-data col-2">–</span>
@@ -233,267 +191,8 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
             <span class="col-data col-1">–</span>
           </div>
         </etools-data-table-row>
-
-        <etools-dialog no-padding keep-dialog-open size="md"
-                ?opened="${this.dialogOpened}"
-                dialog-title="${this.dialogTitle}"
-                .okBtnText="${this.confirmBtnText}"
-                .hideConfirmBtn="${!this.confirmBtnText}"
-                ?show-spinner="${this.requestInProcess}"
-                ?disable-confirm-btn="${this.requestInProcess}"
-                @confirm-btn-clicked="${this._addActionPoint}"
-                openFlag="dialogOpened"
-                @close="${this._resetDialogOpenedFlag}">
-            ${
-              this.notTouched
-                ? html`<div class="copy-warning">It is required to change at least one of the fields below.</div>`
-                : ``
-            }
-                  <div class="container">
-                    <div class="layout-horizontal">
-                        <div class="col col-6">
-                            <!-- Partner -->
-                            <etools-dropdown
-                                    class="${this._setRequired('partner', this.editedApBase)} validate-input fua-person"
-                                    .selected="${this.selectedPartnerIdAux}"
-                                    label="${this.getLabel('partner', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText('partner', this.editedApBase, 'select')}"
-                                    .options="${this.partners}"
-                                    option-label="name"
-                                    option-value="id"
-                                    ?required="${this._setRequired('partner', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly('partner', this.editedApBase, this.requestInProcess)}"
-                                    ?invalid="${this.errors.partner}"
-                                    .errorMessage="${this.errors.partner}"
-                                    @focus="${this._resetFieldError}">
-                            </etools-dropdown>
-                        </div>
-                        <div class="col col-6">
-                            <!-- PD/SSFA -->
-                            <etools-dropdown
-                                    class="${this._setRequired(
-                                      'intervention',
-                                      this.editedApBase
-                                    )} validate-input fua-person"
-                                    .selected="${this.editedItem.intervention?.id}"
-                                    label="${this.getLabel('intervention', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText(
-                                      'intervention',
-                                      this.editedApBase,
-                                      'select'
-                                    )}"
-                                    .options="${this.fullPartner?.interventions}"
-                                    option-label="title"
-                                    option-value="id"
-                                    ?required="${this._setRequired('intervention', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly(
-                                      'intervention',
-                                      this.editedApBase,
-                                      this.requestInProcess
-                                    )}"
-                                    ?invalid="${this.errors.intervention}"
-                                    .errorMessage="${this.errors.intervention}"
-                                    trigger-value-change-event
-                                    @etools-selected-item-changed="${({detail}: CustomEvent) =>
-                                      (this.editedItem.intervention = detail.selectedItem?.id
-                                        ? {id: detail.selectedItem?.id}
-                                        : null)}"
-                                    @focus="${this._resetFieldError}">
-                            </etools-dropdown>
-                        </div>
-                    </div>
-                    <div class="layout-horizontal">
-                        <div class="col col-6">
-                            <!-- Category -->
-                            <etools-dropdown
-                                    class="${this._setRequired(
-                                      'category',
-                                      this.editedApBase
-                                    )} validate-input fua-person"
-                                    .selected="${this.editedItem.category}"
-                                    label="${this.getLabel('category', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText('category', this.editedApBase, 'select')}"
-                                    .options="${this.categories}"
-                                    option-label="display_name"
-                                    option-value="value"
-                                    ?required="${this._setRequired('category', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly('category', this.editedApBase, this.requestInProcess)}"
-                                    ?invalid="${this.errors.category}"
-                                    .errorMessage="${this.errors.category}"
-                                    trigger-value-change-event
-                                    @etools-selected-item-changed="${({detail}: CustomEvent) =>
-                                      (this.editedItem.category = detail.selectedItem?.value)}"
-                                    @focus="${this._resetFieldError}">
-                            </etools-dropdown>
-                        </div>
-                    </div>
-
-                    <div class="layout-horizontal">
-                        <div class="col col-12">
-                            <!-- Description -->
-                            <etools-textarea
-                                    class="w100 validate-input ${this._setRequired('description', this.editedApBase)}"
-                                    .value="${this.editedItem.description}"
-                                    allowed-pattern="[\d\s]"
-                                    label="${this.getLabel('description', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText('description', this.editedApBase)}"
-                                    ?required="${this._setRequired('description', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly(
-                                      'description',
-                                      this.editedApBase,
-                                      this.requestInProcess
-                                    )}"
-                                    max-rows="4"
-                                    ?invalid="${this.errors.description}"
-                                    .errorMessage="${this.errors.description}"
-                                    @value-changed="${({detail}: CustomEvent) =>
-                                      (this.editedItem.description = detail.value)}"
-                                    @focus="${this._resetFieldError}">
-                            </etools-textarea>
-                        </div>
-                    </div>
-
-                    <div class="layout-horizontal">
-                        <div class="col col-6">
-                            <!-- Assigned To -->
-
-                            <etools-dropdown
-                                    class="${this._setRequired(
-                                      'assigned_to',
-                                      this.editedApBase
-                                    )} validate-input fua-person"
-                                    .selected="${this.editedItem.assigned_to?.id}"
-                                    label="${this.getLabel('assigned_to', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText('assigned_to', this.editedApBase, 'select')}"
-                                    .options="${this.users}"
-                                    option-label="name"
-                                    option-value="id"
-                                    .loadDataMethod="${this.loadUsersDropdownOptions}"
-                                    preserve-search-on-close
-                                    ?required="${this._setRequired('assigned_to', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly(
-                                      'assigned_to',
-                                      this.editedApBase,
-                                      this.requestInProcess
-                                    )}"
-                                    ?invalid="${this.errors.assigned_to}"
-                                    .errorMessage="${this.errors.assigned_to}"
-                                    trigger-value-change-event
-                                    @etools-selected-item-changed="${({detail}: CustomEvent) =>
-                                      (this.editedItem = {
-                                        ...this.editedItem,
-                                        assigned_to: {id: detail.selectedItem?.id}
-                                      })}"
-                                    @focus="${this._resetFieldError}">
-                            </etools-dropdown>
-                        </div>
-
-                        <div class="col col-6">
-                            <!-- Sections -->
-
-                            <etools-dropdown
-                                    class="${this._setRequired('section', this.editedApBase)} validate-input fua-person"
-                                    .selected="${this.editedItem.section?.id}"
-                                    label="${this.getLabel('section', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText('section', this.editedApBase, 'select')}"
-                                    .options="${this.sections}"
-                                    option-label="name"
-                                    option-value="id"
-                                    ?required="${this._setRequired('section', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly('section', this.editedApBase, this.requestInProcess)}"
-                                    ?invalid="${this.errors.section}"
-                                    .errorMessage="${this.errors.section}"
-                                    trigger-value-change-event
-                                    @etools-selected-item-changed="${({detail}: CustomEvent) =>
-                                      (this.editedItem = {
-                                        ...this.editedItem,
-                                        section: {id: detail.selectedItem?.id}
-                                      })}"
-                                    @focus="${this._resetFieldError}">
-                            </etools-dropdown>
-                        </div>
-                    </div>
-
-                    <div class="layout-horizontal">
-                        <div class="col col-6">
-                            <!-- Offices -->
-
-                            <etools-dropdown
-                                    class="${this._setRequired('office', this.editedApBase)} validate-input fua-person"
-                                    .selected="${this.editedItem.office?.id}"
-                                    label="${this.getLabel('office', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText('office', this.editedApBase, 'select')}"
-                                    .options="${this.offices}"
-                                    option-label="name"
-                                    option-value="id"
-                                    ?required="${this._setRequired('office', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly('office', this.editedApBase, this.requestInProcess)}"
-                                    ?invalid="${this.errors.office}"
-                                    .errorMessage="${this.errors.office}"
-                                    trigger-value-change-event
-                                    @etools-selected-item-changed="${({detail}: CustomEvent) =>
-                                      (this.editedItem = {
-                                        ...this.editedItem,
-                                        office: {id: detail.selectedItem?.id}
-                                      })}"
-                                    @focus="${this._resetFieldError}">
-                            </etools-dropdown>
-                        </div>
-
-                        <div class="col col-6">
-                            <!-- Due Date -->
-                            <datepicker-lite
-                                    id="deadlineAction"
-                                    class="${this._setRequired('due_date', this.editedApBase)} validate-input"
-                                    .value="${this.editedItem.due_date}"
-                                    label="${this.getLabel('due_date', this.editedApBase)}"
-                                    placeholder="${this.getPlaceholderText('due_date', this.editedApBase, 'select')}"
-                                    ?required="${this._setRequired('due_date', this.editedApBase)}"
-                                    ?readonly="${this.isReadOnly('due_date', this.editedApBase, this.requestInProcess)}"
-                                    ?invalid="${this.errors.due_date}"
-                                    .errorMessage="${this.errors.due_date}"
-                                    @focus="${this._resetFieldError}"
-                                    selected-date-display-format="D MMM YYYY"
-                                    fire-date-has-changed
-                                    @date-has-changed="${(e: CustomEvent) => {
-                                      this.editedItem.due_date = e.detail.date;
-                                    }}"
-                                    >
-                            </datepicker-lite>
-                        </div>
-                    </div>
-
-                    <div class="layout-horizontal">
-                        <!-- High Priority -->
-                        <div class="col col-12 checkbox-container">
-                            <etools-checkbox
-                                    ?checked="${this.editedItem.high_priority}"
-                                    ?disabled="${this.isReadOnly(
-                                      'high_priority',
-                                      this.editedApBase,
-                                      this.requestInProcess
-                                    )}"
-                                    @sl-changed="${(e: any) => (this.editedItem.high_priority = e.target.checked)}">
-                                    This action point is high priority
-                            </etools-checkbox>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="action-complete" ?hidden="${!this._allowComplete(this.editedApBase)}">
-                <etools-button
-                  variant="text"
-                  class="neutral"
-                >
-                    <a href="${this.editedItem.url}" target="_blank">Go To action points to complete
-                        <etools-icon name="launch"></etools-icon>
-                    </a>
-                </etools-button>
-            </div>
-          </div>
-        </etools-dialog>
-            `;
+      </etools-content-panel>
+    `;
   }
 
   @property({type: Array})
@@ -588,19 +287,20 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
   @property({type: Object})
   fullPartner!: any;
 
-  @property({type: Object})
-  loadUsersDropdownOptions?: (search: string, page: number, shownOptionsLimit: number) => void;
-
   public connectedCallback() {
     super.connectedCallback();
 
-    this.loadUsersDropdownOptions = this._loadUsersDropdownOptions.bind(this);
+    this.dialogKey = 'follow-up-actions-dialog';
     this.addEventListener('sort-changed', this._sortOrderChanged as EventListenerOrEventListenerObject);
+    this.addEventListener('show-add-dialog', this.openAddEditDialog as any);
+    this.addEventListener('show-edit-dialog', this.openAddEditDialog as any);
   }
 
   disconnectedCallback() {
     super.disconnectedCallback();
     this.removeEventListener('sort-changed', this._sortOrderChanged as EventListenerOrEventListenerObject);
+    this.addEventListener('show-add-dialog', this.openAddEditDialog as any);
+    this.addEventListener('show-edit-dialog', this.openAddEditDialog as any);
   }
 
   stateChanged(state: RootState) {
@@ -615,9 +315,6 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
   updated(changedProperties: PropertyValues): void {
     super.updated(changedProperties);
 
-    if (changedProperties.has('dialogOpened')) {
-      this._resetDialog(this.dialogOpened);
-    }
     if (changedProperties.has('optionsData')) {
       this.setPermissionPath(this.optionsData);
     }
@@ -627,47 +324,35 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
     if (changedProperties.has('partnerData')) {
       this._requestPartner(this.partnerData);
     }
-    if (changedProperties.has('copyDialog') || changedProperties.has('editedItem')) {
-      this._checkNotTouched(this.copyDialog, this.editedItem);
-    }
   }
 
-  _loadUsersDropdownOptions(search: string, page: number, shownOptionsLimit: number) {
-    const endpoint = clone(famEndpoints.users);
-    endpoint.url += `?page_size=${shownOptionsLimit}&page=${page}&search=${search || ''}`;
-    return sendRequest({
-      method: 'GET',
-      endpoint: {
-        url: endpoint.url
+  openAddEditDialog() {
+    this.dialogOpened = true;
+    openDialog({
+      dialog: this.dialogKey,
+      dialogData: {
+        opener: this,
+        editedItem: this.editedItem,
+        editedApBase: this.editedApBase,
+        users: this.users,
+        sections: this.sections,
+        offices: this.offices,
+        categories: this.categories,
+        partners: this.partners,
+        interventions: this.fullPartner?.interventions || [],
+        selectedPartnerIdAux: this.selectedPartnerIdAux,
+        dialogTitle: this.dialogTitle,
+        confirmBtnText: this.confirmBtnText,
+        copyDialog: this.copyDialog,
+        originalEditedObj: this.originalEditedObj
       }
-    }).then((resp: GenericObject) => {
-      const data = page > 1 ? [...this.users, ...resp.results] : resp.results;
-      this.handleUsersNoLongerAssignedToCurrentCountry(
-        data,
-        this.editedItem.assigned_to ? [this.editedItem.assigned_to] : []
-      );
-      this.users = data;
-    });
-  }
-
-  _allowComplete(editedApBase) {
-    return actionAllowed(editedApBase, 'complete');
+    }).then(() => (this.dialogOpened = false));
   }
 
   _requestPartner(partner) {
     const id = (partner && +partner.id) || null;
     this.partnerId = id;
     this.selectedPartnerId = id;
-  }
-
-  _resetDialog(dialogOpened) {
-    if (dialogOpened) {
-      return;
-    }
-    this.copyDialog = false;
-    this.originalEditedObj = {};
-    this.selectedPartnerIdAux = null;
-    this.resetDialog(dialogOpened);
   }
 
   _sortOrderChanged(e: CustomEvent) {
@@ -738,13 +423,9 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
     return isEmpty(data) ? null : data;
   }
 
-  _addActionPoint() {
-    this._checkNotTouched(this.copyDialog, this.editedItem);
-    if (!this.validate() || this.notTouched) {
-      return;
-    }
-
+  _addActionPoint(editedItem: GenericObject) {
     this.requestInProcess = true;
+    this.editedItem = cloneDeep(editedItem);
     const apData = this.getActionsData();
     if (apData) {
       const method = apData.id ? 'PATCH' : 'POST';
@@ -762,10 +443,12 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
     this.requestInProcess = false;
     if (detail && detail.success) {
       this.dialogOpened = false;
+      this.closeEditDialog();
       if (event.detail.data) {
         this.dataItems = [...event.detail.data];
       }
     } else {
+      this.closeDialogLoading();
       this.errors = event.detail.errors;
     }
   }
@@ -821,22 +504,7 @@ export class FollowUpActions extends connect(store)(CommonMethodsMixin(TableElem
       this.editedItem.assigned_to ? [this.editedItem.assigned_to] : []
     );
     this.users = [...this.users];
-    this.dialogOpened = true;
-  }
-
-  _checkNotTouched(copyDialog, editedItem) {
-    if (!copyDialog || isEmpty(this.originalEditedObj)) {
-      this.notTouched = false;
-      return;
-    }
-    this.notTouched = every(this.originalEditedObj, (value, key) => {
-      const isObj = isObject(value);
-      if (isObj) {
-        return !(value as AnyObject).id || +(value as AnyObject).id === +get(this, `editedItem.${key}.id`);
-      } else {
-        return value === editedItem[key];
-      }
-    });
+    this.openAddEditDialog();
   }
 
   _handleOptionResponse(detail) {
