+++ src_ts/elements/pages/audits/audits-page-main.ts
@@ -20,6 +20,7 @@ import '../../common-elements/follow-up-components/follow-up-main/follow-up-main
 import '../../common-elements/file-attachments-tab/file-attachments-tab';
 import '../../common-elements/status-tab-element/status-tab-element';
 import './report-page-components/audit-report-page-main/audit-report-page-main';
+import '../../common-elements/engagement-cancel/engagement-cancel-dialog';
 import {GenericObject} from '../../../types/global';
 import EngagementMixin from '../../mixins/engagement-mixin';
 import {RootState, store} from '../../../redux/store';
@@ -33,6 +34,7 @@ import get from 'lodash-es/get';
 import {isJsonStrMatch} from '@unicef-polymer/etools-utils/dist/equality-comparisons.util';
 import {fireEvent} from '@unicef-polymer/etools-utils/dist/fire-event.util';
 import {AnyObject} from '@unicef-polymer/etools-types';
+import {openDialog} from '@unicef-polymer/etools-utils/dist/dialog.util';
 
 /**
  * @customElement
@@ -195,38 +197,6 @@ export class AuditsPageMain extends connect(store)(CommonMethodsMixin(Engagement
                 </status-tab-element>
               </div>
             </div>
-
-            <etools-dialog
-              no-padding
-              keep-dialog-open
-              size="md"
-              .opened="${this.dialogOpened}"
-              dialog-title="Cancellation of Engagement"
-              ok-btn-text="Continue"
-              @confirm-btn-clicked="${this._cancelEngagement}"
-              openFlag="dialogOpened"
-              @close="${this._resetDialogOpenedFlag}"
-            >
-              <div class="row-h repeatable-item-container" without-line>
-                <div class="repeatable-item-content">
-                  <div class="row-h group">
-                    <div class="input-container input-container-l">
-                      <etools-textarea
-                        id="cancellationReasonInput"
-                        class="required"
-                        label="Cancellation Reason"
-                        placeholder="Enter reason of cancellation"
-                        required
-                        max-rows="4"
-                        error-message="This field is required."
-                        @focus="${this._resetFieldError}"
-                      >
-                      </etools-textarea>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </etools-dialog>
           `
         : ``}
     `;
@@ -290,6 +260,16 @@ export class AuditsPageMain extends connect(store)(CommonMethodsMixin(Engagement
       return false;
     }
     return true;
+  }
+
+  _openCancelDialog() {
+    openDialog({
+      dialog: 'engagement-cancel-dialog'
+    }).then(({confirmed, response}) => {
+      if (confirmed) {
+        this._cancelEngagement(response);
+      }
+    });
   }
 
   customDataPrepare(data) {
