+++ src_ts/elements/pages/audits/report-page-components/financial-findings/financial-findings.ts
@@ -29,6 +29,7 @@ import {checkNonField} from '../../../../mixins/error-handler';
 import {getTableRowIndexText} from '../../../../utils/utils';
 import {AnyObject} from '@unicef-polymer/etools-utils/dist/types/global.types';
 import '@unicef-polymer/etools-modules-common/dist/layout/are-you-sure';
+import './financial-findings-dialog.js';
 import {openDialog} from '@unicef-polymer/etools-utils/dist/dialog.util';
 
 /**
@@ -159,153 +160,6 @@ export class FinancialFindings extends CommonMethodsMixin(TableElementsMixin(Mod
             <span class="col-data col-1">–</span>
           </div>
         </etools-data-table-row>
-
-        <etools-dialog
-          id="financial-findings"
-          no-padding
-          size="md"
-          ?opened="${this.dialogOpened}"
-          keep-dialog-open
-          dialog-title="${this.dialogTitle}"
-          ok-btn-text="Add"
-          ?show-spinner="${this.requestInProcess}"
-          ?disable-confirm-btn="${this.requestInProcess}"
-          @confirm-btn-clicked="${this._addItemFromDialog}"
-          openFlag="dialogOpened"
-          @close="${this._resetDialogOpenedFlag}"
-        >
-          <div class="layout-horizontal">
-            <div class="col col-6">
-              <!-- Title -->
-              <etools-dropdown
-                id="titleOptionsDropDown"
-                class="w100 ${this._setRequired('financial_finding_set.title', this.optionsData)} validate-input"
-                label="${this.getLabel('financial_finding_set.title', this.optionsData)}"
-                placeholder="${this.getPlaceholderText('financial_finding_set.title', this.optionsData)}"
-                .options="${this.titleOptions}"
-                option-label="display_name"
-                option-value="value"
-                .selected="${this.editedItem.title}"
-                ?required="${this._setRequired('financial_finding_set.title', this.optionsData)}"
-                ?disabled="${this.requestInProcess}"
-                ?invalid="${this.errors.title}"
-                .errorMessage="${this.errors.title}"
-                @focus="${this._resetFieldError}"
-                trigger-value-change-event
-                @etools-selected-item-changed="${({detail}: CustomEvent) =>
-                  this.selectedItemChanged(detail, 'title', 'value', this.editedItem)}"
-                hide-search
-              >
-              </etools-dropdown>
-            </div>
-          </div>
-
-          <div class="layout-horizontal">
-            <div class="col col-6">
-              <!-- Amount (local) -->
-              <etools-currency
-                class="w100 ${this._setRequired('financial_finding_set.local_amount', this.optionsData)} validate-input"
-                .value="${this.editedItem.local_amount}"
-                currency=""
-                label="${this.getLabel('financial_finding_set.local_amount', this.optionsData)}"
-                placeholder="${this.getPlaceholderText('financial_finding_set.local_amount', this.optionsData)}"
-                ?required="${this._setRequired('financial_finding_set.local_amount', this.optionsData)}"
-                ?readonly="${this.requestInProcess}"
-                ?invalid="${this.errors.local_amount}"
-                .errorMessage="${this.errors.local_amount}"
-                @focus="${this._resetFieldError}"
-                @value-changed="${({detail}: CustomEvent) =>
-                  this.numberChanged(detail, 'local_amount', this.editedItem)}"
-              >
-              </etools-currency>
-            </div>
-
-            <div class="col col-6">
-              <!-- Amount USD -->
-              <etools-currency
-                class="w100 ${this._setRequired('financial_finding_set.amount', this.optionsData)} validate-input"
-                .value="${this.editedItem.amount}"
-                currency="$"
-                label="${this.getLabel('financial_finding_set.amount', this.optionsData)}"
-                placeholder="${this.getPlaceholderText('financial_finding_set.amount', this.optionsData)}"
-                ?required="${this._setRequired('financial_finding_set.amount', this.optionsData)}"
-                ?readonly="${this.requestInProcess}"
-                ?invalid="${this.errors.amount}"
-                .errorMessage="${this.errors.amount}"
-                @focus="${this._resetFieldError}"
-                @value-changed="${({detail}: CustomEvent) => this.numberChanged(detail, 'amount', this.editedItem)}"
-              >
-              </etools-currency>
-            </div>
-          </div>
-
-          <div class="layout-horizontal">
-            <div class="col col-12">
-              <!-- Description -->
-              <etools-textarea
-                class="w100 ${this._setRequired('financial_finding_set.description', this.optionsData)}
-                            fixed-width validate-input"
-                .value="${this.editedItem.description}"
-                allowed-pattern="[\\d\\s]"
-                label="${this.getLabel('financial_finding_set.description', this.optionsData)}"
-                placeholder="${this.getPlaceholderText('financial_finding_set.description', this.optionsData)}"
-                ?required="${this._setRequired('financial_finding_set.description', this.optionsData)}"
-                ?disabled="${this.requestInProcess}"
-                max-rows="4"
-                ?invalid="${this.errors.description}"
-                .errorMessage="${this.errors.description}"
-                @focus="${this._resetFieldError}"
-                @value-changed="${({detail}: CustomEvent) => this.valueChanged(detail, 'description', this.editedItem)}"
-              >
-              </etools-textarea>
-            </div>
-          </div>
-
-          <div class="layout-horizontal">
-            <div class="col col-12">
-              <!-- Recommendation -->
-              <etools-textarea
-                class="w100 ${this._setRequired('financial_finding_set.recommendation', this.optionsData)}
-                            fixed-width validate-input"
-                .value="${this.editedItem.recommendation}"
-                allowed-pattern="[\\d\\s]"
-                label="${this.getLabel('financial_finding_set.recommendation', this.optionsData)}"
-                placeholder="${this.getPlaceholderText('financial_finding_set.recommendation', this.optionsData)}"
-                ?required="${this._setRequired('financial_finding_set.recommendation', this.optionsData)}"
-                ?disabled="${this.requestInProcess}"
-                max-rows="4"
-                ?invalid="${this.errors.recommendation}"
-                .errorMessage="${this.errors.recommendation}"
-                @focus="${this._resetFieldError}"
-                @value-changed="${({detail}: CustomEvent) =>
-                  this.valueChanged(detail, 'recommendation', this.editedItem)}"
-              >
-              </etools-textarea>
-            </div>
-          </div>
-
-          <div class="layout-horizontal">
-            <div class="col col-12">
-              <!-- IP comments -->
-              <etools-textarea
-                class="w100 ${this._setRequired('financial_finding_set.ip_comments', this.optionsData)}
-                            fixed-width validate-input"
-                .value="${this.editedItem.ip_comments}"
-                allowed-pattern="[\\d\\s]"
-                label="${this.getLabel('financial_finding_set.ip_comments', this.optionsData)}"
-                placeholder="${this.getPlaceholderText('financial_finding_set.ip_comments', this.optionsData)}"
-                ?required="${this._setRequired('financial_finding_set.ip_comments', this.optionsData)}"
-                ?disabled="${this.requestInProcess}"
-                max-rows="4"
-                ?invalid="${this.errors.ip_comments}"
-                .errorMessage="${this.errors.ip_comments}"
-                @focus="${this._resetFieldError}"
-                @value-changed="${({detail}: CustomEvent) => this.valueChanged(detail, 'ip_comments', this.editedItem)}"
-              >
-              </etools-textarea>
-            </div>
-          </div>
-        </etools-dialog>
       </etools-content-panel>
     `;
   }
@@ -347,12 +201,17 @@ export class FinancialFindings extends CommonMethodsMixin(TableElementsMixin(Mod
 
   connectedCallback() {
     super.connectedCallback();
+    this.dialogKey = 'financial-findings-dialog';
     this.addEventListener('show-confirm-dialog', this.openConfirmDeleteDialog as any);
+    this.addEventListener('show-add-dialog', this.openAddEditDialog as any);
+    this.addEventListener('show-edit-dialog', this.openAddEditDialog as any);
   }
 
   disconnectedCallback() {
     super.disconnectedCallback();
     this.removeEventListener('show-confirm-dialog', this.openConfirmDeleteDialog as any);
+    this.addEventListener('show-add-dialog', this.openAddEditDialog as any);
+    this.addEventListener('show-edit-dialog', this.openAddEditDialog as any);
   }
 
   updated(changedProperties: PropertyValues): void {
@@ -368,6 +227,19 @@ export class FinancialFindings extends CommonMethodsMixin(TableElementsMixin(Mod
     if (changedProperties.has('optionsData')) {
       this.setChoices(this.optionsData);
     }
+  }
+
+  openAddEditDialog() {
+    openDialog({
+      dialog: this.dialogKey,
+      dialogData: {
+        optionsData: this.optionsData,
+        editedItem: this.editedItem,
+        opener: this,
+        dialogTitle: this.dialogTitle,
+        titleOptions: this.titleOptions
+      }
+    });
   }
 
   openConfirmDeleteDialog() {
