+++ src_ts/elements/pages/special-audits/special-audits-page-main.ts
@@ -17,6 +17,7 @@ import '../../data-elements/update-engagement';
 import './report-page-components/sa-report-page-main/sa-report-page-main';
 import '../../common-elements/file-attachments-tab/file-attachments-tab';
 import '../../common-elements/pages-header-element/pages-header-element';
+import '../../common-elements/engagement-cancel/engagement-cancel-dialog';
 
 import EngagementMixin from '../../mixins/engagement-mixin';
 import CommonMethodsMixin from '../../mixins/common-methods-mixin';
@@ -35,6 +36,7 @@ import get from 'lodash-es/get';
 import {fireEvent} from '@unicef-polymer/etools-utils/dist/fire-event.util';
 import {tabInputsStyles} from '../../styles/tab-inputs-styles';
 import {AnyObject} from '@unicef-polymer/etools-types';
+import {openDialog} from '@unicef-polymer/etools-utils/dist/dialog.util';
 
 /**
  * @customElement
@@ -208,38 +210,6 @@ export class SpecialAuditsPageMain extends connect(store)(CommonMethodsMixin(Eng
                 </status-tab-element>
               </div>
             </div>
-
-            <etools-dialog
-              no-padding
-              keep-dialog-open
-              size="md"
-              .opened="${this.dialogOpened}"
-              dialog-title="Cancellation of Engagement"
-              ok-btn-text="Continue"
-              @confirm-btn-clicked="${this._cancelEngagement}"
-              openFlag="dialogOpened"
-              @close="${this._resetDialogOpenedFlag}"
-            >
-              <div class="row-h repeatable-item-container" without-line>
-                <div class="repeatable-item-content">
-                  <div class="row-h group">
-                    <div class="input-container input-container-l">
-                      <etools-textarea
-                        id="cancellationReasonInput"
-                        class="required"
-                        label="Cancellation Reason"
-                        placeholder="Enter reason of cancellation"
-                        required
-                        max-rows="4"
-                        error-message="This field is required."
-                        @focus="${this._resetFieldError}"
-                      >
-                      </etools-textarea>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </etools-dialog>
           `
         : ``}
     `;
@@ -346,5 +316,15 @@ export class SpecialAuditsPageMain extends connect(store)(CommonMethodsMixin(Eng
       return false;
     }
     return true;
+  }
+
+  _openCancelDialog() {
+    openDialog({
+      dialog: 'engagement-cancel-dialog'
+    }).then(({confirmed, response}) => {
+      if (confirmed) {
+        this._cancelEngagement(response);
+      }
+    });
   }
 }
